🌸 update: move logger to util
