+++ app.ts
@@ -1,24 +1,21 @@
 import createError, { HttpError } from 'http-errors';
 import express, { json, NextFunction, Request, Response, urlencoded } from 'express';
 import cookieParser from 'cookie-parser';
-import uniqueValidator from 'mongoose-unique-validator';
 import mongoose from 'mongoose';
+import cors from 'cors';
+import 'module-alias/register';
+
+import uniqueValidator from 'mongoose-unique-validator';
 import mongooseAutoPopulate from 'mongoose-autopopulate';
 import mongoosePaginate from 'mongoose-paginate-v2';
 import aggregatePaginate from 'mongoose-aggregate-paginate-v2';
-import morgan from 'morgan';
-import fs from 'fs';
-import { createStream, Generator, Options } from 'rotating-file-stream';
-import path from 'path';
-import 'module-alias/register';
-import cors from 'cors';
-import dayjs from 'dayjs';
 
 import indexRouter from '@routes/indexRouter';
 import accountRouter from '@routes/accountRouter';
 import seedRouter from '@routes/seedRouter';
-import { dateTimeNowTz } from '@utils/dateTimeUtil';
+
 import createRestAPIError, { RestAPIError } from '@error/createRestAPIError';
+import { accessLogger, errorLogger, seedLogger } from '@utils/logUtil';
 
 const app = express();
 
@@ -59,40 +56,9 @@ app.use(cookieParser());
 //---------------------
 //   LOGGER
 //---------------------
-const format = ':date[Asia/Bangkok] == :remote-user :method :url :status :response-time ms == :res[content-length]';
-
-morgan.token('date', (_req, _res, tz) => {
-  return dateTimeNowTz(tz as string).format('ddd, DD MMM YYYY HH:mm:ss');
-});
-
-const logName: (name: string) => Generator = (name) => (time: number | Date, index: number | undefined) => {
-  if (!time) return path.resolve(__dirname, 'log', name, `${name}.log`);
-  return path.resolve(__dirname, 'log', name, `${dayjs().format('YYYY-MMDD-HHmm')}-${index}-${name}.log`);
-};
-
-const logOpt: Options = { omitExtension: true, interval: '7d', maxFiles: 4, history: 'log-history.txt' };
-
-if (!fs.existsSync('log')) fs.mkdirSync('log', { recursive: true });
-
-app.use(
-  morgan(format, {
-    skip: (_req, res) => res.statusCode < 400,
-    stream: createStream(logName('error'), logOpt),
-  })
-);
-
-app.use(
-  morgan(format, {
-    skip: (req, _res) => req.baseUrl !== '/seed',
-    stream: createStream(logName('seed'), logOpt),
-  })
-);
-
-app.use(
-  morgan(format, {
-    stream: createStream(logName('access'), logOpt),
-  })
-);
+app.use(errorLogger);
+app.use(seedLogger);
+app.use(accessLogger);
 
 //---------------------
 //   ROUTES
